+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [UNRELEASE]
 
 - Fix missing images in exported Knowledge Base PDFs
+- escape non-HTML tags in PDF output
 
 ## [4.0.1] - 2025-03-10